refactor(app): isolate startup construction
